bit of code clean up
